feat(shapewidget): add a way to change text alignment and position
This commit adds a way to choose how the text is aligned vertically by changing the value of
textProps.verticalAlign. Currently, the possible values are TOP, BOTTOM and MIDDLE (defined in
SVGLandmarkRepresentation/Constants). The ShapeWidget example was updated to match this change.

Another change concerns the positionning of the text inside the shape bounding box.
The `computeTextPosition` function was updated to use an array of 3 elements indicating
the position on the axis X, Y and Z. The allowed values are TextPosition.MIN, TextPosition.CENTER
and TextPosition.MAX. EllipseWidget and RectangleWidget states have been updated to match
this change.